Linear / Angular receiving Floats instead of Doubles for easy dealing with Loomo SDK

+++ app/src/main/java/com/kiwicampus/loomo/models/freedom_command/Angular.kt
@@ -3,5 +3,5 @@ package com.kiwicampus.loomo.models.freedom_command
 data class Angular(
     val x: Double,
     val y: Double,
-    val z: Double
+    val z: Float
 )
